Move datastore init in StorageModule
